Get base hidden assignment view to work; TODO: keybinds, unhide button, transitions for hiding/unhiding, rename view hidden to view deleted, rename icon to recover, api.views logic for marking as hidden instead of deleting

+++ timeweb/timewebapp/static/timewebapp/js/utils.js
@@ -1152,14 +1152,16 @@ for (let sa of dat) {
 };
 // Use DOMContentLoaded because $(function() { fires too slowly
 document.addEventListener("DOMContentLoaded", function() {
-    utils.reloadAtMidnight();
-    if (SETTINGS.gc_integration_enabled) ajaxUtils.createGCAssignments();
-    utils.ui.setClickHandlers.tickButtons();
-    utils.ui.setClickHandlers.assignmentsHeaderUI();
-    utils.ui.setClickHandlers.assignmentSorting();
-    utils.ui.setClickHandlers.deleteAllStarredAssignments();
-    utils.ui.setClickHandlers.deleteAssignmentsFromClass();
-    utils.ui.setClickHandlers.autofillWorkDone();
+    if (!VIEW_HIDDEN_ASSIGNMENTS) {
+        utils.reloadAtMidnight();
+        if (SETTINGS.gc_integration_enabled) ajaxUtils.createGCAssignments();
+        utils.ui.setClickHandlers.tickButtons();
+        utils.ui.setClickHandlers.assignmentsHeaderUI();
+        utils.ui.setClickHandlers.assignmentSorting();
+        utils.ui.setClickHandlers.deleteAllStarredAssignments();
+        utils.ui.setClickHandlers.deleteAssignmentsFromClass();
+        utils.ui.setClickHandlers.autofillWorkDone();
+    }
     utils.ui.addTagHandlers();
     utils.ui.setKeybinds();
     utils.ui.displayFullDueDateOnHover();
